+++ src/main/java/mod/icarus/crimsonrevelations/item/armor/CRItemMeteorBoots.java
@@ -47,7 +47,7 @@ public class CRItemMeteorBoots extends ItemArmor implements ISpecialArmor, IRech
                 boost = (float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_LAND_SPEED;
                 return player.isSneaking() ? boost / 4.0F : boost;
             case WATER_GROUND:
-                boost = (float) Math.max((float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_LAND_SPEED / (float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_SNEAK_REDUCTION,
+                boost = Math.max((float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_LAND_SPEED / (float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_SNEAK_REDUCTION,
                         (float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_WATER_SPEED);
                 return player.isSneaking() ? boost / (float) CRConfig.thaumic_augmentation_settings.METEOR_BOOTS_SNEAK_REDUCTION : boost;
             case WATER_SWIM:
@@ -188,8 +188,8 @@ public class CRItemMeteorBoots extends ItemArmor implements ISpecialArmor, IRech
         // Particles when sprinting or jumping
         if (!world.isRemote) {
             if (!player.isInWater() && (motion > 0.1F || player.isSprinting())) {
-                ((WorldServer) world).spawnParticle(EnumParticleTypes.FLAME, (double) (player.posX + Math.random() - 0.5F),
-                        (double) (player.getEntityBoundingBox().minY + 0.25F + ((Math.random() - 0.5) * 0.25F)), (double) (player.posZ + Math.random() - 0.5F), 1, 0.0D, 0.025D, 0.0D, 0.0D);
+                ((WorldServer) world).spawnParticle(EnumParticleTypes.FLAME, player.posX + Math.random() - 0.5F,
+                        player.getEntityBoundingBox().minY + 0.25F + ((Math.random() - 0.5) * 0.25F), player.posZ + Math.random() - 0.5F, 1, 0.0D, 0.025D, 0.0D, 0.0D);
             }
         }
